requirements.txt
flask
pyrebase4
firebase_admin
flask_cors

azure-cognitiveservices-vision-computervision
python-dotenv
PIL